invalidate fetcher cache if hash files are different
SNAPSHOT builds, either local or from the artifacts API, are expected to be different even though their names are the same. 

Therefore this change makes the fetcher cache to fetch the hash file and compare with the cached hash file only for SNAPSHOT versions. Non-SNAPSHOT versions are considered to be stable.